backend/requirements.txt
Django~=5.0.2
djangorestframework~=3.14.0
requests==2.26.0
python-dotenv==1.0.1
